requirements.txt
reflex~=0.8.28.post1
pandas~=2.3.3
supabase~=2.28.3
python-dotenv~=1.2.2
psycopg2-binary~=2.9.11
bcrypt~=5.0.0
google-cloud-bigquery~=3.40.1
db-dtypes~=1.5.0
langgraph~=1.1.8
langchain-groq~=1.1.2
markdown-pdf~=1.13.1
PyPDF2~=3.0.1
